requirements.txt
neo4j
pandas
pyarrow
python-dotenv
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-openai
openai>=1.0.0
tiktoken
numpy
